requirements.txt
Flask==1.1.2
Flask-Session==0.3.2
Flask-Sessions==0.1.5
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.3
greenlet==0.4.16
itsdangerous==1.1.0
Jinja2==2.11.2
jsmin==2.2.2
MarkupSafe==1.1.1
monotonic==1.5
nodeenv==1.4.0
PyMySQL==0.9.3
python-dotenv==0.14.0
python-engineio==3.13.1
python-socketio==4.6.0
redis==3.5.3
six==1.15.0
SQLAlchemy==1.3.18
webassets==2.0
Werkzeug==1.0.1
cachelib==0.1.1
click==7.1.2
dnspython==1.16.0
eventlet==0.26.1
